Stop the apps stating money rules the platform does not implement
Every figure the platform charges or requires is a `Platform_Settings` row.
Five screens across the three apps stated their own, and four of the five
were wrong.

**The customer could not withdraw money the platform would have paid.**
`BottleWallet.tsx` carried

    /** Mirrors the server-side minimums so the UI rejects before the round trip. */
    const MIN_TOP_UP_KSH = 10;
    const MIN_WITHDRAWAL_KSH = 500;

and the comment was false. `settlement_service.withdrawal_terms` returns a
minimum of **1** and a fee of **0** for a customer — unspent wallet credit
is their own money coming back, not earnings. The screen refused every
withdrawal under KSH 500, client-side, before any request was sent, so no
server log recorded a refusal because none happened. The rider and vendor
wallets had this removed already; the customer's was missed because there
was no customer wallet-summary to read the terms from. There is now: the
terms ride on `GET /api/bottle-returns/summary`, which already served that
screen's balance. `topup.minimum` joins the rider's and vendor's summaries
for the same reason.

**A wholesale vendor was told the retail minimum.** "Minimum amount is
KSH 500" is `payout_min_retail_vendor`; theirs is 1,000. The figure was in
the copy, and `summary.withdrawal.minimum` was already on the screen.

**The customer was warned about the wrong cancellation fee, or one they
would not be charged.** "a KSH 50 cancellation penalty will apply" is the
pre-pickup row; after pickup it is 150, and it is waived entirely while a
free cancellation remains in the rolling 30 days. The app knows none of
those three things, so it no longer states a number — and
`cancel_customer_order`, which was adding up to KSH 150 to `debt_balance`
and returning "Order cancelled successfully", now reports
`penalty_charged` and `free_cancellations_remaining`. The customer is told
what they were charged instead of meeting it as a line on their next order.

**The vendor app typed every wire money field as `number`.**
`vendor_wallet_summary` serialises all of them through `money_str`, so
`tsc` was checking that screen's whole money path against a shape the
server has never sent. The rider app's copy of the interface had it right.

**Money on the way *in* was `float`.** `TopUpRequest.amount`,
`WithdrawRequest.amount`, `ConfigPreview.product_total`, `deposit_fee` and
`min_order_value` — the last being what `assert_meets_minimum` compares
against a `Decimal` subtotal on every checkout. `test_money_serialisation`
could not see any of them: it walked `schemas/` only, and all five are
declared beside the routes that use them. It now walks anything declaring
a `BaseModel`, scoped to Pydantic fields rather than bare annotations, and
`rounding_adjustment` and `amount` join `MONEY_FIELDS`. The customer's
top-up and withdrawal now cross the wire as decimal strings; `Number()` is
gone from that screen entirely.

`test_inline_money_math.py` gains two rules: no business money value as a
named constant, and none written into a sentence. Both proven against a
`git archive HEAD` snapshot — 1 and 5 failures respectively.

Backend 2951 · customer 152 · vendor 119 · rider 129 · four typechecks clean

+++ BackendAPI/routes/wallet_routes.py
@@ -2,6 +2,8 @@ import logging
 
 from fastapi import APIRouter, Depends, Header, Request, HTTPException, Query
 from fastapi.responses import JSONResponse
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -23,7 +25,13 @@ router = APIRouter(prefix="/api/wallet", tags=["Wallet"])
 
 
 class TopUpRequest(BaseModel):
-    amount: float = Field(gt=0, le=150_000)
+    # `Decimal`, not `float`. Money annotated `float` on a Pydantic model is the
+    # quiet version of a `float(...)` cast — there is nothing to grep for,
+    # because Pydantic does the coercion. It is the *inbound* half of the rule
+    # the schemas already follow, and it was invisible to
+    # `test_money_serialisation.py` because that walk only ever looked at
+    # `schemas/`, and these two models are declared in a route file.
+    amount: Decimal = Field(gt=0, le=150_000)
     phone_number: str
     # Which of the caller's wallets to act on. Verified against the token in
     # `resolve_wallet_owner` — a client cannot name an account it does not own.
@@ -31,7 +39,7 @@ class TopUpRequest(BaseModel):
 
 
 class WithdrawRequest(BaseModel):
-    amount: float = Field(gt=0, le=150_000)
+    amount: Decimal = Field(gt=0, le=150_000)
     phone_number: str
     user_type: str = "customer"
